Update faq.md
Misc. updates - linking the rby manip guide several times. Updating links to guides to send them to the correct place. Added Blue NSC, since it is actually a great category to increase exposure to manips. Explanation that we cannot link bios file. Included GSE as an approved emulator - link to SRC page since that is where the most current version of emus can be found.